@@ -2,8 +2,4 @@
  * @typedef {import('../packages/icons-scripts/scripts/options').DeprecatedIcons} DeprecatedIcons
  * @type DeprecatedIcons
  */
-const DEPRECATED_ICONS = {};
-
-module.exports = {
-  DEPRECATED_ICONS,
-};
+export const DEPRECATED_ICONS = {};
