@@ -1,7 +1,2 @@
-const { generateIcons } = require('./scripts/icons');
-const { createIconsMap } = require('./scripts/icons-map');
-
-module.exports = {
-  generateIcons,
-  createIconsMap,
-};
+export { generateIcons } from './scripts/icons.js';
+export { createIconsMap } from './scripts/icons-map.js';
